pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# This is a scripts-only project, not a Python package
packages = []

[project]
name = "claude-trading-skills"
version = "0.1.0"
description = "Claude Skills for equity investors and traders"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.9.0",
    "bandit[toml]>=1.7.0",
    "detect-secrets>=1.5.0",
    "codespell>=2.3.0",
    "pre-commit>=3.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = [
    "E501",  # line-too-long (handled by formatter)
    "E402",  # module-level import not at top (sys.path manipulation in conftest)
    "B904",  # raise-without-from-inside-except (fix incrementally)
    "B007",  # unused-loop-control-variable (fix incrementally)
    "B017",  # assert-raises-exception (fix incrementally)
]

[tool.ruff.lint.per-file-ignores]
"**/conftest.py" = ["E402"]
"**/test_*.py" = ["E402"]

[tool.pytest.ini_options]
# Tests must be run per-skill to avoid module name collisions
# (each skill has its own scorer.py, calculators/, helpers.py).
# Use scripts/run_all_tests.sh for safe per-skill execution.
# The testpaths below are used by `pytest` for quick single-invocation runs.
testpaths = [
    "skills/backtest-expert/scripts/tests",
    "skills/canslim-screener/tests",
    "skills/dual-axis-skill-reviewer/scripts/tests",
    "skills/ftd-detector/scripts/tests",
    "skills/institutional-flow-tracker/scripts/tests",
    "skills/macro-regime-detector/scripts/tests",
    "skills/market-breadth-analyzer/scripts/tests",
    "skills/market-top-detector/scripts/tests",
    "skills/stanley-druckenmiller-investment/scripts/tests",
    "skills/theme-detector/scripts/tests",
    "skills/theme-detector/tests",
    "skills/uptrend-analyzer/scripts/tests",
    "skills/vcp-screener/scripts/tests",
    "skills/earnings-trade-analyzer/scripts/tests",
    "skills/pead-screener/scripts/tests",
    "skills/position-sizer/scripts/tests",
    "skills/data-quality-checker/scripts/tests",
    "skills/edge-strategy-reviewer/scripts/tests",
    "skills/edge-pipeline-orchestrator/scripts/tests",
    "skills/skill-idea-miner/scripts/tests",
]

[tool.bandit]
targets = ["skills"]
skips = [
    "B101",  # assert used (normal in non-production code)
    "B110",  # try-except-pass
    "B113",  # requests without timeout (pre-existing in test scripts)
    "B310",  # urllib urlopen (used for FMP API calls with validated URLs)
    "B311",  # random (used for non-security skill selection)
    "B404",  # import subprocess (needed for test runner)
    "B603",  # subprocess without shell=true (intentionally safe pattern)
]

[tool.codespell]
skip = ".venv,*.json,*.csv,*.baseline,skill-packages,*.pyc,__pycache__"
ignore-words-list = "trough,ore,cna,acn,blok,hims,pont,claus"

[tool.coverage.run]
source = ["skills"]

[tool.coverage.report]
fail_under = 40
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
]
